requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pdfplumber>=0.10.0
python-docx>=0.8.11
Pillow>=10.0.0
pytesseract>=0.3.10
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
httpx>=0.25.0
python-dotenv>=1.0.0
